v1.1 Added nginx, and compose.yaml instant hot-reloading
The nginx issue was when a client goes to /Dashboard, nginx returns a 404, unlike with `serve` command nginx doesnt know that when user goes to /Dashboard to serve the react app.

The database.db wasn't being created inside the Container, because the Mac direct path doesn't work on Linux. Which is what the container runs on.

Fixed hot-reloading for both the ./logs and the databade.db to show changes that are happening inside the container on the host machine.